Decoder module
It turned out that I was doing endian-ness in Instruction.cpp very wrong. I have no idea why I thought that was the way to do it. This is also now fixed.

As part of this mistake, there was no need at all for any decoding to occur in the alu. This will be addressed in the next commit. Note: for the time being, alu tests are broken.